complete refactoring, split logic from commands
